requirements.txt
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ollama==0.6.0
packaging==25.0
pdf2image==1.17.0
pillow==12.0.0
playwright==1.55.0
pydantic==2.12.3
pydantic_core==2.41.4
pyee==13.0.0
pytesseract==0.3.13
sniffio==1.3.1
typing-inspection==0.4.2
typing_extensions==4.15.0
